delete message in group chat & friend chat reset token, notifications, group icon & friend image

+++ lib/features/groups/data/services/group_firebase_services.dart
@@ -2,7 +2,6 @@ import 'dart:io';
 
 import 'package:chat_app/features/groups/data/model/group_data.dart';
 import 'package:chat_app/features/groups/data/model/group_message_data.dart';
-import 'package:chat_app/helper/notification_services.dart';
 import 'package:chat_app/utils/constants.dart';
 import 'package:chat_app/utils/data/models/user.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -10,7 +9,6 @@ import 'package:firebase_auth/firebase_auth.dart' hide User;
 import 'package:firebase_storage/firebase_storage.dart';
 
 class GroupFirebaseServices {
-  final _notificationServices = NotificationServices();
   final FirebaseStorage _storage = FirebaseStorage.instance;
   final _usersCollection =
       FirebaseFirestore.instance.collection(FirebasePath.users);
@@ -18,32 +16,61 @@ class GroupFirebaseServices {
       FirebaseFirestore.instance.collection(FirebasePath.groups);
 
   Stream<List<Group>> getAllUserGroups() {
+    final currentUser = FirebaseAuth.instance.currentUser;
+    if (currentUser == null) {
+      return Stream.value([]);
+    }
+
     return _usersCollection
-        .doc(FirebaseAuth.instance.currentUser!.uid)
-        .collection(FirebasePath.groups)
+        .doc(currentUser.uid)
         .snapshots()
-        .map(
-          (querySnapshot) => querySnapshot.docs
-              .map(
-                (queryDocSnapshot) => Group.fromJson(queryDocSnapshot.data()),
-              )
-              .toList(),
-        );
+        .asyncMap((snapshot) async {
+      final dynamic userData = snapshot.data();
+      if (userData != null && userData is Map<String, dynamic>) {
+        final List<dynamic> groupIds =
+            (userData['groups'] ?? []) as List<dynamic>;
+        final List<Future<Group>> groupFutures = groupIds.map((groupId) async {
+          final DocumentSnapshot groupSnapshot =
+              await _groupsCollection.doc(groupId.toString()).get();
+          return Group.fromJson(groupSnapshot.data()! as Map<String, dynamic>);
+        }).toList();
+        final List<Group> groups = await Future.wait(groupFutures);
+        return groups;
+      } else {
+        return [];
+      }
+    });
   }
 
   Future<List<User>> getAllGroupMembers(String groupId) async {
     final DocumentSnapshot<Map<String, dynamic>> groupSnapshot =
         await _groupsCollection.doc(groupId).get();
     if (groupSnapshot.exists) {
-      final List<dynamic> memberDataList =
+      final List<dynamic> memberIds =
           groupSnapshot.data()?['members'] as List<dynamic>;
-      final List<User> users = memberDataList.map((userData) {
-        return User.fromJson(userData as Map<String, dynamic>);
-      }).toList();
+      final List<User> users = [];
+      for (final memberId in memberIds) {
+        final DocumentSnapshot userSnapshot =
+            await _usersCollection.doc(memberId.toString()).get();
+        if (userSnapshot.exists) {
+          final userData = userSnapshot.data();
+          if (userData != null && userData is Map<String, dynamic>) {
+            final User user = User.fromJson(userData);
+            users.add(user);
+          }
+        }
+      }
       return users;
     } else {
       return [];
     }
+  }
+
+  Future<String> getAdminName(String adminId) async {
+    final DocumentSnapshot<Map<String, dynamic>> adminSnapshot =
+        await _usersCollection.doc(adminId).get();
+    final adminData = adminSnapshot.data();
+    return adminData?['userName'] as String? ?? 'Unknown';
   }
 
   Stream<List<GroupMessage>> getAllGroupMessages(String groupId) {
@@ -62,18 +89,7 @@ class GroupFirebaseServices {
         );
   }
 
-  static CollectionReference<Group> getGroupsCollection() {
-    return FirebaseFirestore.instance
-        .collection(FirebasePath.users)
-        .doc(FirebaseAuth.instance.currentUser!.uid)
-        .collection(FirebasePath.groups)
-        .withConverter<Group>(
-          fromFirestore: (snapshot, _) => Group.fromJson(snapshot.data()!),
-          toFirestore: (equipment, options) => equipment.toJson(),
-        );
-  }
-
-  Stream<List<Group>> getGroups() {
+  Stream<List<Group>> getGroupsForSearch() {
     return _groupsCollection.snapshots().map(
           (querySnapshot) => querySnapshot.docs
               .map(
@@ -87,38 +103,35 @@ class GroupFirebaseServices {
     Group group,
     User currentUser,
   ) async {
-    final groups = getGroupsCollection();
-    final userGroupDocRef = await groups.add(group);
+    final userGroupDocRef = await _groupsCollection.add(group.toJson());
     group.groupId = userGroupDocRef.id;
-    final Map<String, dynamic> currentUserMap = currentUser.toJson();
-    _groupsCollection.doc(group.groupId).set(group.toJson());
+
+    final String currentUserId = currentUser.id!;
     final groupDocRef = _groupsCollection.doc(group.groupId);
     groupDocRef.update({
       "members": FieldValue.arrayUnion([
-        currentUserMap,
+        currentUserId,
       ]),
     });
     await userGroupDocRef.update({
       "members": FieldValue.arrayUnion([
-        currentUserMap,
+        currentUserId,
       ]),
       "groupId": group.groupId,
     });
-    await _usersCollection.doc(FirebaseAuth.instance.currentUser!.uid).update({
+    await _usersCollection.doc(currentUserId).update({
       "groups": FieldValue.arrayUnion([group.groupId]),
     });
   }
 
   Future<void> joinGroup(Group group, User user) async {
-    _usersCollection
-        .doc(FirebaseAuth.instance.currentUser!.uid)
-        .collection(FirebasePath.groups)
-        .doc(group.groupId)
-        .set(group.toJson());
+    final String currentUserId = FirebaseAuth.instance.currentUser!.uid;
+
     await _groupsCollection.doc(group.groupId).update({
-      "members": FieldValue.arrayUnion([user.toJson()]),
+      "members": FieldValue.arrayUnion([currentUserId]),
     });
-    await _usersCollection.doc(FirebaseAuth.instance.currentUser!.uid).update({
+
+    await _usersCollection.doc(currentUserId).update({
       "groups": FieldValue.arrayUnion([group.groupId]),
     });
   }
@@ -129,30 +142,25 @@ class GroupFirebaseServices {
         storageRef.child('${imageFile.hashCode}').putFile(imageFile);
     final TaskSnapshot snapshot = await uploadImage;
     final String downloadUrl = await snapshot.ref.getDownloadURL();
-   return downloadUrl;
-  }
-  Future<String> uploadImageAndUpdateGroupIcon(File imageFile, String groupId) async {
-      final String downloadUrl = await uploadImage(imageFile);
-      await _groupsCollection.doc(groupId).update({
-        'groupIcon': downloadUrl,
-      });
-      final QuerySnapshot usersSnapshot = await _usersCollection.get();
-      for (final userDoc in usersSnapshot.docs) {
-        final String userId = userDoc.id;
-        final DocumentReference userGroupDocRef = _usersCollection
-            .doc(userId)
-            .collection(FirebasePath.groups)
-            .doc(groupId);
-        final DocumentSnapshot userGroupSnapshot = await userGroupDocRef.get();
-        if (userGroupSnapshot.exists) {
-          await userGroupDocRef.update({
-            'groupIcon': downloadUrl,
-          });
-        }
-      }
-      return downloadUrl;
+    return downloadUrl;
   }
 
+  Future<String> uploadImageAndUpdateGroupIcon(
+    File imageFile,
+    String groupId,
+  ) async {
+    final String downloadUrl = await uploadImage(imageFile);
+    await _groupsCollection.doc(groupId).update({
+      'groupIcon': downloadUrl,
+    });
+    return downloadUrl;
+  }
+
+  Future<User> getUserData(String userId) async {
+    final currentUserId = userId;
+    final docSnapshot = await _usersCollection.doc(currentUserId).get();
+    return User.fromJson(docSnapshot.data()!);
+  }
 
   Future<void> sendMessageToGroup(
     Group group,
@@ -196,36 +204,19 @@ class GroupFirebaseServices {
       'recentMessage': message,
       'recentMessageSender': sender.userName,
     });
-
-    _usersCollection
-        .doc(currentUserUid)
-        .collection(FirebasePath.groups)
-        .doc(group.groupId)
-        .update({
-      'recentMessage': message,
-      'recentMessageSender': sender.userName,
-    });
-    final List<String> memberIds =
-        group.members!.map((member) => member.id!).toList();
-    for (final memberId in memberIds) {
-      await _notificationServices.sendNotification(
-        fcmToken: group.members
-                ?.firstWhere((member) => member.id == memberId)
-                .fCMToken ??
-            '',
-        title: 'New Message in ${group.groupName}',
-        body: message,
-      );
-    }
   }
 
-  Stream<bool> isUserInGroup(String groupId) {
-    return _usersCollection
-        .doc(FirebaseAuth.instance.currentUser!.uid)
-        .collection(FirebasePath.groups)
-        .doc(groupId)
-        .snapshots()
-        .map((event) => event.exists);
+  Stream<bool> isUserInGroup(String groupId, String userId) {
+    return _groupsCollection.doc(groupId).snapshots().map((snapshot) {
+      if (snapshot.exists) {
+        final Map<String, dynamic>? data = snapshot.data();
+        if (data != null && data.containsKey('members')) {
+          final List<dynamic> members = data['members'] as List<dynamic>;
+          return members.contains(userId);
+        }
+      }
+      return false;
+    });
   }
 
   Future<void> leaveGroup(Group group, User user) async {
@@ -238,13 +229,26 @@ class GroupFirebaseServices {
       "groups": FieldValue.arrayRemove([group.groupId]),
     });
     await _groupsCollection.doc(group.groupId).update({
-      "members": FieldValue.arrayRemove([user.toJson()]),
+      "members": FieldValue.arrayRemove([user.id]),
     }).whenComplete(() async {
-      final groupSnapshot = await _groupsCollection.doc(group.groupId).get();
-      final groupData = Group.fromJson(groupSnapshot.data()!);
-      if (groupData.members!.isEmpty && groupData.adminId == user.id) {
-        await _groupsCollection.doc(group.groupId).delete();
+      if (group.members!.isEmpty) {
+        if (user.id == group.adminId) {
+          final groupSnapshot =
+              await _groupsCollection.doc(group.groupId).get();
+          final groupData = Group.fromJson(groupSnapshot.data()!);
+          if (groupData.members!.isEmpty && groupData.adminId == user.id) {
+            await _groupsCollection.doc(group.groupId).delete();
+          }
+        }
       }
     });
+  }
+
+  Future<void> deleteMessageForeAll(String groupId, String messageId) async {
+    _groupsCollection
+        .doc(groupId)
+        .collection(FirebasePath.messages)
+        .doc(messageId)
+        .delete();
   }
 }
